Restructure project into package and scripts
