+++ README.md
@@ -1,6 +1,9 @@
 # EDGAR 10-K Fetcher
 
-This repository provides simple Python utilities for working with the SEC EDGAR system.  It includes a script to download the latest 10‑K filing and another to list the files available for that filing.  A third helper retrieves the master list of company CIKs and names.
+This repository provides Python utilities for working with the SEC EDGAR system.
+It now exposes a package in `edgar/` and command-line tools under `scripts/`.
+The tools can download the latest 10‑K filing, list its files, fetch the company
+CIK list, and monitor EDGAR for new filings while uploading them to S3.
 
 All scripts throttle requests to roughly **six per second** to remain within the SEC rate limits.
 
@@ -27,7 +30,7 @@ If `tqdm` is missing, progress bars are disabled but the scripts still run.
 Run the script with a company CIK (Central Index Key):
 
 ```bash
-python edgar_fetcher.py <CIK>
+python scripts/fetch_10k.py <CIK>
 ```
 
 The latest 10‑K document will be saved in the `10k/` directory. Replace `<CIK>` with a valid CIK number (e.g., Apple Inc. is `0000320193`).
@@ -35,7 +38,7 @@ The latest 10‑K document will be saved in the `10k/` directory. Replace `<CIK>
 To list the files available in the latest 10‑K filing:
 
 ```bash
-python edgar_files.py <CIK>
+python scripts/list_files.py <CIK>
 ```
 This prints each file name along with its description, form type, and size.
 
@@ -44,7 +47,7 @@ Use `--json-out <path>` to write the list to a JSON file instead of printing it.
 To download a list of all company CIKs and names:
 
 ```bash
-python edgar_companies.py > companies.txt
+python scripts/companies.py > companies.txt
 ```
 
 The script outputs each CIK and company name on a single line, which can be
@@ -53,17 +56,17 @@ redirected to a file for later reference.
 ## Example
 
 ```bash
-python edgar_fetcher.py 0000320193
+python scripts/fetch_10k.py 0000320193
 ```
 
 This command downloads the most recent 10‑K for Apple and stores it locally.
 
 ## Monitoring and S3 Upload
 
-Use `edgar_monitor.py` to check for new filings and upload their documents to an S3 bucket.
+Use `scripts/monitor.py` to check for new filings and upload their documents to an S3 bucket.
 
 ```bash
-python edgar_monitor.py <CIK> [<CIK> ...] --bucket <bucket-name> [--prefix path/] \
+python scripts/monitor.py <CIK> [<CIK> ...] --bucket <bucket-name> [--prefix path/] \
                          [--state state.json] [--manifest manifest.json]
 ```
 
@@ -77,5 +80,5 @@ finishes. This serves as a catalog of everything retrieved so far.
 Example:
 
 ```bash
-python edgar_monitor.py 0000320193 --bucket my-bucket --manifest manifests/apple.json
+python scripts/monitor.py 0000320193 --bucket my-bucket --manifest manifests/apple.json
 ```
