+++ edgar/companies.py
@@ -1,12 +1,13 @@
+"""Helpers for retrieving company tickers list."""
 from typing import List, Dict
 
-from edgar_fetcher import sec_get
+from .client import sec_get
 
 COMPANY_TICKERS_URL = "https://www.sec.gov/files/company_tickers.json"
 
 
 def fetch_cik_company_list() -> List[Dict[str, str]]:
-    """Download list of companies with CIK and name."""
+    """Download mapping of CIK to company names."""
     resp = sec_get(COMPANY_TICKERS_URL)
     resp.raise_for_status()
     data = resp.json()
@@ -16,12 +17,3 @@ def fetch_cik_company_list() -> List[Dict[str, str]]:
         name = entry.get('title', '').title()
         companies.append({"cik": cik, "name": name})
     return companies
-
-
-if __name__ == "__main__":
-    comps = fetch_cik_company_list()
-    try:
-        for c in comps:
-            print(c["cik"], c["name"])
-    except BrokenPipeError:
-        pass
